analytics/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.24.0
pandas==2.0.0
scipy==1.11.0
scikit-learn==1.3.0
scikit-image==0.21.0
nltk==3.8.1
textblob==0.17.1
transformers==4.33.0
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
tensorflow==2.13.0
keras==2.13.0
spacy==3.7.0
gensim==4.3.0
fasttext==0.9.2
sentence-transformers==2.2.2
huggingface-hub==0.17.0
pillow==10.0.0
opencv-python==4.8.0
requests==2.31.0
gunicorn==21.2.0
pydantic==2.4.0
redis==5.0.0
celery==5.3.0
python-magic==0.4.27
librosa==0.10.0
pydub==0.25.1
diffusers==0.21.0
accelerate==0.24.0
safetensors==0.4.0
pytorch-lightning==2.1.0
optuna==3.14.0
tensorboard==2.14.0
wandb==0.16.0
